+++ go.mod
@@ -2,8 +2,14 @@ module github.com/stytchauth/stytch-connected-apps-cli-example
 
 go 1.22.1
 
+require github.com/zalando/go-keyring v0.2.6
+
 require (
+	al.essio.dev/pkg/shellescape v1.5.1 // indirect
+	github.com/danieljoos/wincred v1.2.2 // indirect
+	github.com/godbus/dbus/v5 v5.1.0 // indirect
 	github.com/inconshreveable/mousetrap v1.1.0 // indirect
 	github.com/spf13/cobra v1.9.1 // indirect
 	github.com/spf13/pflag v1.0.6 // indirect
+	golang.org/x/sys v0.26.0 // indirect
 )
